isValid
if the word is horizontal, check if the vertical words are valid. check
above for letter until a black space, and then get the start location,
and check below the word. put the word and location into object results
and into a string array. if it's only one letter (not word, = valid).
then return Results